[유치수] [NOBTS] int를 short으로 강제변환하도록 임시로 수정. 변수 타입을 전부 바꾸는게 좋을 듯.
git-svn-id: http://svn.bds.nhncorp.com/pe/hippo-web/trunk@3255 84d0f5b1-2673-498c-a247-62c4ff18d310

+++ src/main/java/com/nhn/pinpoint/web/vo/LinkStatistics.java
@@ -130,7 +130,8 @@ public class LinkStatistics {
 			SortedMap<Long, Long> map = timeseriesValueList.get(i);
 
 			// 다른 slot에도 같은 시간이 존재해야한다.
-			if (i == timeseriesSlotIndex.get(responseTimeslot)) {
+			// FIXME responseTimeSlot의 자료형을 short으로 변경할 것.
+			if (i == timeseriesSlotIndex.get((short) responseTimeslot)) {
 				long v = map.containsKey(timestamp) ? map.get(timestamp) + callCount : callCount;
 				map.put(timestamp, v);
 			} else {
